backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# Video processing
opencv-python-headless==4.9.0.80
yt-dlp==2024.5.27

# AI / Embeddings
# torch and torchvision are installed separately in Dockerfile
# from PyTorch's own CDN for reliability and speed
transformers==4.41.2
Pillow==10.3.0

# Vector search
faiss-cpu==1.8.0

# Utilities
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
